auto-model-deploy/requirements.txt
aws-cdk.aws-events
aws-cdk.aws-events-targets
aws-cdk.aws-lambda
aws-cdk.cdk
aws-cdk.aws-iam
aws-cdk.aws-s3
boto3